+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* support for the broken pipe error in stdin writing
 
+### Changed
+* `rust-version = "1.58.0"`
 
 ## [0.2.9] (2023-02-12)
 ### Removed